+++ src/io/gro_io.rs
@@ -7,9 +7,12 @@ use std::fs::File;
 use std::io::{BufRead, BufReader, BufWriter, Write};
 use std::path::Path;
 
+use crate::aux::{GRO_MAX_COORDINATE, GRO_MIN_COORDINATE};
 use crate::errors::{ParseGroError, WriteGroError};
 use crate::structures::{atom::Atom, simbox::SimBox};
 use crate::system::System;
+
+use super::check_coordinate_sizes;
 
 /// ## Methods for writing gro files.
 impl System {
@@ -84,6 +87,16 @@ impl System {
     ) -> Result<(), WriteGroError> {
         if !self.group_exists(group_name) {
             return Err(WriteGroError::GroupNotFound(group_name.to_string()));
+        }
+
+        // check that coordinates of the atoms are in the range supported by the data format
+        // this has to be done before the file is even created
+        if !check_coordinate_sizes(
+            self.group_iter(group_name).unwrap(),
+            GRO_MIN_COORDINATE,
+            GRO_MAX_COORDINATE,
+        ) {
+            return Err(WriteGroError::CoordinateTooLarge);
         }
 
         let output = File::create(&filename)
@@ -791,5 +804,31 @@ mod tests_write {
         let mut expected = File::open("test_files/example_box_zero.gro").unwrap();
 
         assert!(file_diff::diff_files(&mut result, &mut expected));
+    }
+
+    #[test]
+    fn write_too_large_coordinate_1() {
+        let mut system = System::from_file("test_files/example.gro").unwrap();
+
+        system.get_atom_as_mut(16).unwrap().set_position_z(10000.0);
+
+        match system.write_gro("will_not_be_created_1.gro", false) {
+            Err(WriteGroError::CoordinateTooLarge) => (),
+            Ok(_) => panic!("Writing should have failed, but it did not."),
+            Err(e) => panic!("Incorrect error type `{:?}` was returned.", e),
+        }
+    }
+
+    #[test]
+    fn write_too_large_coordinate_2() {
+        let mut system = System::from_file("test_files/example.gro").unwrap();
+
+        system.get_atom_as_mut(16).unwrap().set_position_x(-9999.0);
+
+        match system.group_write_gro("all", "will_not_be_created_2.gro", false) {
+            Err(WriteGroError::CoordinateTooLarge) => (),
+            Ok(_) => panic!("Writing should have failed, but it did not."),
+            Err(e) => panic!("Incorrect error type `{:?}` was returned.", e),
+        }
     }
 }