+++ changelog.md
@@ -37,6 +37,7 @@
 - Implemented `System::traj_iter_map_reduce` for simple embarrassingly parallel iteration through simulation trajectories.
 - Implemented `System::group_intersection` allowing to directly create groups that are intersections of other groups.
 - Implemented `System::from_file_with_format` allowing to directly specify the format of the input file.
+- When writing 'gro' and 'pdb' files, large coordinates can no longer 'overflow' the space allocated for them in the file format and instead an error is returned.
 
 ***